splitting new_confirmations on several modules to reduce the module size
